--- src/Pyroclast/model/stokes_2D_mg/__init__.py
@@ -1,1 +1,0 @@
-from .model import IncompressibleStokes2DMG
