--- src/Pyroclast/solvers/multigrid/__init__.py
@@ -1,5 +1,0 @@
-"""Multigrid solver infrastructure for Pyroclast."""
-
-from .base import BaseMultigrid
-
-__all__ = ["BaseMultigrid"]
